fix port in use and improve tests files
