+++ __tests__/controllers/auth/create-account.spec.ts
@@ -1,0 +1,31 @@
+import request from 'supertest';
+import app from '@app';
+import { factory } from 'typeorm-seeding';
+import { User } from '@entities/user.entity';
+import { API } from '../../utils';
+
+describe('creating an account', () => {
+  it('returns http code 200 whith valid params', async () => {
+    const userFields = await factory(User)().make();
+    const response = await request(app)
+      .post(`${API}/auth/signup`)
+      .send(userFields);
+    expect(response.status).toBe(200);
+  });
+
+  it('returns http code 400 whith invalid params', async () => {
+    const userFields = {};
+    const response = await request(app)
+      .post(`${API}/auth/signup`)
+      .send(userFields);
+    expect(response.status).toBe(400);
+    expect(response.body).toStrictEqual(
+      expect.objectContaining({
+        errMessage: expect.any(String),
+        errCode: expect.any(Number)
+      })
+    );
+  });
+});
+
+
