+++ __tests__/controllers/users/create-user.spec.ts
@@ -1,0 +1,17 @@
+import request from 'supertest';
+import { getRepository } from 'typeorm';
+import { factory } from 'typeorm-seeding';
+import app from '@app';
+import { User } from '@entities/user.entity';
+import { API } from '../../utils';
+describe('creating a user', () => {
+  it('returns http code 200 and creates the user', async () => {
+    const userFields = await factory(User)().make();
+
+    const response = await request(app).post(`${API}/users`).send(userFields);
+    expect(response.status).toBe(200);
+
+    const userRepo = getRepository<User>(User);
+    expect(await userRepo.count()).toBeGreaterThan(0);
+  });
+});
