+++ src/server.ts
@@ -4,7 +4,7 @@ import AdminBro from 'admin-bro';
 import AdminBroExpress from '@admin-bro/express';
 import { Database, Resource } from '@admin-bro/typeorm';
 import { User } from '@entities/user.entity';
-import { PORT, TYPEORM_PORT, REDIS_PORT } from '@config';
+import { PORT, TYPEORM_PORT, REDIS_PORT, TESTING_ENV } from '@config';
 import connection from '@database/connection';
 import app from '@app';
 import { createRedisClient } from '@utils/redis/client';
@@ -35,13 +35,15 @@ const handleConnection = async (connection: Connection) => {
   app.use(adminBro.options.rootPath, router);
 
   // run express application on given port
-  app.listen(PORT, () => {
-    return console.log(
-      `Server is listening on ${PORT}
+  if (!TESTING_ENV) {
+    app.listen(PORT, () => {
+      return console.log(
+        `Server is listening on ${PORT}
 Postgres database is listening on port ${TYPEORM_PORT}
 Redis database is listening on port ${REDIS_PORT}`
-    );
-  });
+      );
+    });
+  }
 };
 
 // Start redis server
